# Stop copying LogicalPlan and Exprs in `PushDownFilter` (4%-6% faster planning)